requirements.txt
virtualenv
requests
python-dotenv
pandas
psycopg2-binary